Criação do componente de carreira e inserção do item carreira no menu
