+++ src/components/SocialMedia.jsx
@@ -2,6 +2,7 @@ import linkedinIcon from "../assets/images/social-media/linkedin.png";
 import emailIcon from "../assets/images/social-media/email.png";
 import whatsappIcon from "../assets/images/social-media/whatsapp.png";
 import githubIcon from "../assets/images/social-media/github.png";
+import { motion } from "framer-motion";
 
 const socialLinks = [
   {
